Import client JS from single app source
