+++ src/javascripts/_deprecated/company-edit.js
@@ -1,0 +1,42 @@
+/* eslint no-new: 0 */
+const { hide, show } = require('../../lib/element-stuff')
+const Sectors = require('./sectors')
+require('./lookup-address')
+
+const CompanyEdit = {
+  init () {
+    const tradingAddressWrapper = document.getElementById('trading-address-wrapper')
+    const showTradingButton = document.getElementById('add-trading-address')
+    const hideTradingButton = document.getElementById('remove-trading-address')
+
+    if (!showTradingButton || !hideTradingButton || !tradingAddressWrapper) { return }
+
+    const tradingAddressFields = tradingAddressWrapper.querySelectorAll('input, select')
+    const tradingAddressCountryField = tradingAddressWrapper.querySelector('[name=trading_address_country]')
+    const registeredCountryField = document.querySelector('[name=registered_address_country]')
+    const registeredCountryId = (registeredCountryField) ? registeredCountryField.value : null
+
+    showTradingButton.addEventListener('click', function (event) {
+      event.preventDefault()
+      tradingAddressCountryField.value = registeredCountryId
+      hide(showTradingButton)
+      show(hideTradingButton)
+      show(tradingAddressWrapper)
+    }, true)
+
+    hideTradingButton.addEventListener('click', function (event) {
+      event.preventDefault()
+      show(showTradingButton)
+      hide(hideTradingButton)
+      hide(tradingAddressWrapper)
+
+      for (let index = 0; index < tradingAddressFields.length; index += 1) {
+        tradingAddressFields.item(index).value = ''
+      }
+    }, true)
+
+    Sectors.init()
+  },
+}
+
+module.exports = CompanyEdit
