--- src/javascripts/service-delivery.js
@@ -1,1 +1,0 @@
-require('./sectors')
