+++ src/views/company/_layout-view.njk
@@ -27,8 +27,3 @@
 {% block main_grid_right_column %}
   {% include "_messages.njk" %}
 {% endblock %}
-
-{% block javascript %}
-  {{ super() }}
-  <script src="/javascripts/archive-form.bundle.js"></script>
-{% endblock %}
