+++ src/views/company/add-step-1.njk
@@ -59,7 +59,3 @@
     </div>
   </form>
 {% endblock %}
-
-{% block javascript_edit %}
-  <script src="/javascripts/company-add.bundle.js"></script>
-{% endblock %}
