+++ src/views/company/contacts.njk
@@ -79,8 +79,3 @@
     </ul>
   {% endif %}
 {% endblock %}
-
-{% block javascript %}
-  {{ super() }}
-  <script src="/javascripts/expandable-card.bundle.js"></script>
-{% endblock %}
