+++ src/views/company/exports-edit.njk
@@ -61,7 +61,3 @@
     {{ trade.save(backUrl = "/company-exports/view/" + company.id) }}
   </form>
 {% endblock %}
-
-{% block javascript_edit %}
-  <script src="/javascripts/add-another-field.bundle.js"></script>
-{% endblock %}
