+++ src/views/company/interactions.njk
@@ -82,8 +82,3 @@
     <p id="no-interaction-warning" class="infostrip">You currently have no interactions for this company.</p>
   {% endif %}
 {% endblock %}
-
-{% block javascript %}
-  {{ super() }}
-  <script src="/javascripts/expandable-card.bundle.js"></script>
-{% endblock %}
