+++ src/views/contact/edit.njk
@@ -103,8 +103,3 @@
     {% endif %}
   </form>
 {% endblock %}
-
-{% block javascript %}
-  {{ super() }}
-  <script src="/javascripts/contact-edit.bundle.js"></script>
-{% endblock %}
