+++ src/views/contact/interactions.njk
@@ -58,8 +58,3 @@
     <a id="add-interaction-link" class="button button-secondary" href="{{addInteractionUrl}}">Add interaction</a>
   </p>
 {% endblock %}
-
-{% block javascript %}
-  {{ super() }}
-  <script src="/javascripts/expandable-card.bundle.js"></script>
-{% endblock %}
