+++ webpack.config.js
@@ -4,17 +4,9 @@ const path = require('path')
 const isProd = process.env.NODE_ENV === 'production'
 
 const webpackConfig = {
-  devtool: 'source=map',
+  devtool: 'source-map',
   entry: {
     app: './src/javascripts/app.js',
-    'company-add': './src/javascripts/company-add',
-    'company-edit': './src/javascripts/company-edit',
-    'contact-edit': './src/javascripts/contact-edit',
-    'expandable-card': './src/javascripts/expandable-card',
-    index: './src/javascripts/index',
-    'service-delivery': './src/javascripts/service-delivery',
-    'archive-form': './src/javascripts/archive-form',
-    'add-another-field': './src/javascripts/add-another-field',
     ie: ['html5shiv'],
   },
   module: {
